Terminal state is now alwyas correct
If OpenKit is in TimeSync state when it's shut down the shutdown state
has to decided based on whether it's an initial time sync or not.
If it's an initial time sync -> no data collected so far, it's ok
to go to the Terminal state immediately.
If it's not an initial time sync there might be data available which
needs to be transferred to the server (flushed).