Fix wrong pattern for SimpleDateFormat
In Java 6 the pattern string X is unknown for SimpleDateFormat.
This would just add the timezone in ISO-8601 format to the string.
Due to the fact that the DefaultLogger uses UTC timezone, this does
not add useful information in the string. Therefore it's removed.

Also did refactor the tests and some documentation.